feat: rename file with kebab-case
