+++ src/router/calories-counter-router.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { NavigationContainer, } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
-import { AddFood } from '../screens/AddFood'
-import { CaloriesCounter } from '../screens/CaloriesCounter'
+import { AddFood } from '../screens/add-food'
+import { CaloriesCounter } from '../screens/calories-counter'
 
 type RootStackParamList = {
   'Home': undefined
